move all olm functions to olm.go
